Refactor inductor_lowering.py into two files
stack-info: PR: https://github.com/pytorch/helion/pull/1103, branch: jansel/stack/223
